fixup! build: work correctly when linking to a system llvm which is not the default
The names of the llvm components to request --libs flags for, was
accidentally removed in the refactoring. Add it back.

By default llvm-config will assume "all", and for the standard shared
library case llvm will only ever link as one large shared library.

But when JULIACODEGEN is not "llvm", and static builds (or llvm built as
many smaller shared libraries, which is un-recommended except for llvm
developers) are requested this will result in overlinking.

The compiler is supposed to optimize this out by only pulling in the
symbols it needs, and shared library builds can always use
-Wl,--as-needed, but it is "more proper" to do this correctly from the
start...